created class , add for teacher file and correct file input
